Updates
-Finished sudoku solver. Basically it can solve N x N Sudoku puzzles.
I implemented it in a generic way so every size should be supported.

Of course this is not finished.
Later on I will try to add GUI via Swing (never done this, so we'll see how it goes).

In addition, I will add more puzzles with different sizes, I want it to support puzzles size of to 25 x 25.